pyproject.toml
[project]
name = "aicte"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "groq",
    "pypdf>=6.6.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.2",
]
